WIP [REACTOS] Fix warning C4146: unary minus operator applied to unsigned type, result still unsigned
Also make it a critical warning

+++ dll/win32/msi/action.c
@@ -2413,7 +2413,7 @@ static BYTE *parse_value( MSIPACKAGE *package, const WCHAR *value, DWORD len, DW
                 p++;
             }
             if (deformated[0] == '-')
-                d = -d;
+                d = -(LONG)d;
             *(DWORD *)data = d;
             TRACE( "DWORD %lu\n", *(DWORD *)data);
